Update Windows source bundle generation.

+++ source/pidgin/win32/generate_source_bundle.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # Pidgin++ Source Bundles Generator
-# Copyright (C) 2014, 2015 Renato Silva
+# Copyright (C) 2014-2016 Renato Silva
 # Licensed under GNU GPLv2 or later
 
 test -z "$3" && exit 1
@@ -23,13 +23,17 @@ for package in "${packages[@]}"; do
     package_version=$(package_version ${package_name})
     source_name=$(package_source ${package})
     source_package="${source_name}-${package_version}.src.tar.gz"
-    echo "Integrating ${source_package}"
-    for file in "${source_name}"*src.tar.gz; do
-        [[ "${file}" = "${source_package}" && -s "${file}" ]] && continue
+    source_package_simple="${source_package/-${package_architecture}/}"
+    echo "Integrating ${source_package_simple}"
+    for file in "${source_name}"*src.tar.gz "${source_name/-${package_architecture}/}"*src.tar.gz; do
+        [[ "${file}" = "${source_package}"        && -s "${file}" ]] && continue
+        [[ "${file}" = "${source_package_simple}" && -s "${file}" ]] && continue
         rm -f "${file}"
     done
-    [[ -s "${source_package}" ]] && continue
-    if ! wget --quiet "http://repo.msys2.org/mingw/sources/${source_package}"; then
+    [[ -s "${source_package}"        ]] && continue
+    [[ -s "${source_package_simple}" ]] && continue
+    if ! wget --quiet "http://repo.msys2.org/mingw/sources/${source_package}" &&
+       ! wget --quiet "http://repo.msys2.org/mingw/sources/${source_package_simple}"; then
         warn "failed downloading ${source_package}"
         echo "${source_package}" >> MISSING.txt
         rm -f "${source_package}"
